+++ src/Models/GratitudeShoutout.php
@@ -15,7 +15,6 @@ use App\Plugins\PluginModel;
  * @property string $message
  * @property string|null $category
  * @property int|null $points_awarded
- * @property string|null $posted_channel_id
  * @property \Illuminate\Support\Carbon $created_at
  * @property \Illuminate\Support\Carbon $updated_at
  */
@@ -30,7 +29,6 @@ class GratitudeShoutout extends PluginModel
         'message',
         'category',
         'points_awarded',
-        'posted_channel_id',
     ];
 
     /** @var array<string, string> */